+++ CHANGELOG.md
@@ -13,11 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-*
+* Ensures result in ASGI response is a sequence
 
 ### Fixed
 
-*
+* String conversion in ASGI response
 
 ## [1.35.0] - 2026-01-06